Restructuing the project:
1. 'main' scripts is now outside dirs;

2. DATA dir is now inside the project;

3. Dirs with scripts are renamed wih respect to their tasks.
