+++ .gitattributes
@@ -1,0 +1,2 @@
+*.tar filter=lfs diff=lfs merge=lfs -text
+DATA/data_raw/*.tar filter=lfs diff=lfs merge=lfs -text
