+++ DATA/data_raw/LC08_L2SP_172016_20200709_20200912_02_T1_EXAMPLE.tar
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d52c4ab90c262e7257074131d7639db653e262704843b83c247e3ee3ac45a8ab
+size 76772692
